+++ packages/astro-when/tests/fixture/hybrid-output/src/pages/on-demand.ts
@@ -1,0 +1,8 @@
+import { whenAmI } from '@it-astro:when';
+import type { APIRoute } from 'astro';
+
+export const prerender = false;
+
+export const GET: APIRoute = () => {
+	return new Response(whenAmI);
+};