+++ packages/content-utils/src/integration/index.ts
@@ -1,12 +1,14 @@
 import { withApi, onHook, registerGlobalHooks } from '@inox-tools/modular-station';
-import { emptyState } from './state.js';
+import { emptyState, type IntegrationState } from './state.js';
 import { resolveContentPaths } from '../internal/resolver.js';
-import { mkdirSync, writeFileSync } from 'node:fs';
+import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
 import { addVitePlugin, defineIntegration } from 'astro-integration-kit';
 import { injectorPlugin } from './injectorPlugin.js';
 import { seedCollections, type SeedCollectionsOptions } from './seedCollections.js';
 import { gitBuildPlugin, gitDevPlugin } from './gitPlugin.js';
 import { debug } from '../internal/debug.js';
+import * as devalue from 'devalue';
+import { z } from 'astro/zod';
 
 export type InjectCollectionOptions = {
 	/**
@@ -27,9 +29,16 @@ export type InjectCollectionOptions = {
 export const integration = withApi(
 	defineIntegration({
 		name: '@inox-tools/content-utils',
-		setup: () => {
+		optionsSchema: z
+			.object({
+				staticOnlyCollections: z.array(z.string()).optional().default([]),
+			})
+			.optional()
+			.default({}),
+		setup: ({ options: { staticOnlyCollections } }) => {
 			debug('Generating empty state');
 			const state = emptyState();
+			state.staticOnlyCollections.push(...staticOnlyCollections);
 			const collectionSeedBuffer: SeedCollectionsOptions[] = [];
 
 			const api = {
@@ -99,12 +108,56 @@ export const integration = withApi(
 							seedCollections(state, seedOptions);
 						}
 					},
+					'astro:build:done': async () => {
+						await clearStaticCollections(state);
+						for (const cleanup of state.cleanups) {
+							await cleanup();
+						}
+					},
 				},
 				...api,
 			};
 		},
 	})
 );
+
+async function clearStaticCollections(state: IntegrationState) {
+	// After the build is done, if there was such a chunk and there are collections
+	// that should only be present during static build, clean them.
+	if (
+		!(
+			state.staticOnlyCollections.length > 0 &&
+			state.contentDataEntrypoint &&
+			existsSync(state.contentDataEntrypoint)
+		)
+	)
+		return;
+
+	const originalContent = readFileSync(state.contentDataEntrypoint, 'utf-8');
+
+	// Content was already cleared by Astro. Collections are not used anywhere on server bundle
+	if (!originalContent.includes('export')) return;
+
+	// Import the chunk, which exports a devalue flattened map as the default export
+	const { default: value } = await import(/*@vite-ignore*/ state.contentDataEntrypoint);
+
+	// Unflatten the map
+	const map: Map<string, Map<string, unknown>> = devalue.unflatten(value);
+
+	// Remove all the collections we promise we won't use on the server
+	for (const collection of state.staticOnlyCollections) {
+		map.delete(collection);
+	}
+
+	// Build the source code with the new map flattened
+	const newContent = [
+		`const _astro_dataLayerContent = ${devalue.stringify(map)}`,
+		'\nexport { _astro_dataLayerContent as default }',
+	].join('\n');
+
+	// Write it back
+	writeFileSync(state.contentDataEntrypoint, newContent, 'utf-8');
+}
 
 const triggers = Symbol.for('@inox-tools/content-utils:triggers/gitTrackedListResolved');