+++ packages/content-utils/src/runtime/fancyContent.ts
@@ -1,19 +1,16 @@
 /// <reference types="astro/client" />
 
-import { defineCollection as defineNative, z, type SchemaContext } from 'astro:content';
+import {
+	defineCollection as defineNative,
+	z,
+	type BaseSchema,
+	type SchemaContext,
+} from 'astro:content';
+
+export type { BaseSchema };
 
 // Copied from Starlight because Astro core doesn't export their definitions :(
 // https://github.com/withastro/starlight/blob/95ace6d51c7bf631863b06ee46bbc9b6dff6828a/packages/starlight/schema.ts#L116-L123
-
-/** Plain object, union, and intersection Zod types. */
-type BaseSchemaWithoutEffects =
-	| z.AnyZodObject
-	| z.ZodUnion<[BaseSchemaWithoutEffects, ...BaseSchemaWithoutEffects[]]>
-	| z.ZodDiscriminatedUnion<string, z.AnyZodObject[]>
-	| z.ZodIntersection<BaseSchemaWithoutEffects, BaseSchemaWithoutEffects>;
-
-/** Base subset of Zod types that we support passing to the `extend` option. */
-export type BaseSchema = BaseSchemaWithoutEffects | z.ZodEffects<BaseSchemaWithoutEffects>;
 
 /** Type that extends a collection's default schema with an optional, user-defined schema. */
 export type ExtendedSchema<S extends BaseSchema, T extends BaseSchema | never = never> = [