+++ packages/content-utils/tests/fixture/static-collections/astro.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'astro/config';
+import contentUtils from '@inox-tools/content-utils';
+
+export default defineConfig({
+	integrations: [
+		contentUtils({
+			staticOnlyCollections: ['static-only'],
+		}),
+	],
+});