+++ packages/server-islands/src/runtime/debug.ts
@@ -1,0 +1,7 @@
+import debugC from 'debug';
+
+export const debug = debugC('inox-tools:server-islands');
+
+export const getDebug = (segment?: string) => {
+	return segment ? debug.extend(segment) : debug;
+};