+++ packages/server-islands/virtual.d.ts
@@ -1,0 +1,3 @@
+declare module '@it-astro:server-islands' {
+	export * from '@inox-tools/server-islands/runtime/lib';
+}