+++ packages/aik-route-config/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @inox-tools/aik-route-config
 
+## 0.9.1
+
+### Patch Changes
+
+- Updated dependencies [10fe460]
+- Updated dependencies [1a1687b]
+  - @inox-tools/utils@0.4.0
+
+## 0.9.0
+
+### Minor Changes
+
+- 5392c77: Bump Astro version
+
+## 0.8.2
+
+### Patch Changes
+
+- Updated dependencies [094efca]
+  - @inox-tools/utils@0.3.1
+
 ## 0.8.1
 
 ### Patch Changes