+++ packages/astro-when/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inox-tools/astro-when
 
+## 1.1.0
+
+### Minor Changes
+
+- 5392c77: Bump Astro version
+
+## 1.0.3
+
+### Patch Changes
+
+- 3818a2f: Fixes projects with `output: 'static'` being flagged as `StaticOutput` even when an adapter is present.
+
+  Previously, using this was the explicit `output: 'hybrid'`, now this mode is set by the presence of an adapter.
+
 ## 1.0.2
 
 ### Patch Changes