+++ packages/modular-station/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @inox-tools/modular-station
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [10fe460]
+- Updated dependencies [1a1687b]
+  - @inox-tools/utils@0.4.0
+
+## 0.5.0
+
+### Minor Changes
+
+- 5392c77: Bump Astro version
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [094efca]
+  - @inox-tools/utils@0.3.1
+
 ## 0.4.1
 
 ### Patch Changes