+++ packages/request-nanostores/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @inox-tools/request-nanostores
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies [10fe460]
+- Updated dependencies [1a1687b]
+  - @inox-tools/utils@0.4.0
+  - @inox-tools/request-state@0.4.1
+
+## 0.3.0
+
+### Minor Changes
+
+- 5392c77: Bump Astro version
+- 5392c77: Bump Nanostores to v1
+
+### Patch Changes
+
+- Updated dependencies [5392c77]
+  - @inox-tools/request-state@0.4.0
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [094efca]
+  - @inox-tools/utils@0.3.1
+  - @inox-tools/request-state@0.3.2
+
 ## 0.2.2
 
 ### Patch Changes