+++ packages/runtime-logger/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @inox-tools/runtime-logger
 
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [10fe460]
+- Updated dependencies [1a1687b]
+  - @inox-tools/utils@0.4.0
+  - @inox-tools/modular-station@0.5.1
+
+## 0.5.0
+
+### Minor Changes
+
+- 5392c77: Bump Astro version
+
+### Patch Changes
+
+- Updated dependencies [5392c77]
+  - @inox-tools/modular-station@0.5.0
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [094efca]
+  - @inox-tools/utils@0.3.1
+  - @inox-tools/modular-station@0.4.2
+
 ## 0.4.1
 
 ### Patch Changes